+++ MobileGarage/viewmodel/src/commonTest/kotlin/com/chriscartland/garage/viewmodel/RealNetworkButtonHealthRepositoryPropagationTest.kt
@@ -32,6 +32,7 @@ import com.chriscartland.garage.domain.model.User
 import com.chriscartland.garage.testcommon.FakeAppLoggerRepository
 import com.chriscartland.garage.testcommon.FakeAuthRepository
 import com.chriscartland.garage.testcommon.FakeDiagnosticsCountersRepository
+import com.chriscartland.garage.testcommon.FakeDoorCommandRepository
 import com.chriscartland.garage.testcommon.FakeDoorFcmRepository
 import com.chriscartland.garage.testcommon.FakeDoorRepository
 import com.chriscartland.garage.testcommon.FakeFeatureAllowlistRepository
@@ -41,6 +42,7 @@ import com.chriscartland.garage.testcommon.FakeRemoteButtonRepository
 import com.chriscartland.garage.testcommon.FakeStatusSnapshotStore
 import com.chriscartland.garage.testcommon.TestDispatcherProvider
 import com.chriscartland.garage.usecase.ButtonHealthDisplay
+import com.chriscartland.garage.usecase.CheckDoorCommandUseCase
 import com.chriscartland.garage.usecase.CheckInStalenessManager
 import com.chriscartland.garage.usecase.ClassifyVoiceIntentUseCase
 import com.chriscartland.garage.usecase.ComputeButtonHealthDisplayUseCase
@@ -171,6 +173,7 @@ class RealNetworkButtonHealthRepositoryPropagationTest {
             deregisterFcmUseCase = DeregisterFcmUseCase(FakeDoorFcmRepository()),
             signInWithGoogleUseCase = SignInWithGoogleUseCase(authRepo),
             pushRemoteButtonUseCase = PushRemoteButtonUseCase(authRepo, FakeRemoteButtonRepository()),
+            checkDoorCommandUseCase = CheckDoorCommandUseCase(authRepo, FakeDoorCommandRepository()),
             checkInStalenessManager = stalenessManager,
             liveClock = liveClock,
             buttonHealthDisplay = computeButtonHealth(),